Adds in radio button patch request and text for page
EDUCATOR-229
