+++ lms/static/js/discussions_management/models/course_discussions_settings.js
@@ -6,7 +6,8 @@
             defaults: {
                 divided_inline_discussions: [],
                 divided_course_wide_discussions: [],
-                always_divide_inline_discussions: false
+                always_divide_inline_discussions: false,
+                division_scheme: 'none'
             }
         });
         return CourseDiscussionsSettingsModel;
