+++ lms/static/js/spec/groups/views/discussions_spec.js
@@ -20,22 +20,26 @@ define(['backbone', 'jquery', 'edx-ui-toolkit/js/utils/spec-helpers/ajax-helpers
 
             createMockDiscussionsSettingsJson = function(dividedInlineDiscussions,
                                                          dividedCourseWideDiscussions,
-                                                         alwaysDivideInlineDiscussions) {
+                                                         alwaysDivideInlineDiscussions,
+                                                         availableDivisionSchemes) {
                 return {
                     id: 0,
                     divided_inline_discussions: dividedInlineDiscussions || [],
                     divided_course_wide_discussions: dividedCourseWideDiscussions || [],
-                    always_divide_inline_discussions: alwaysDivideInlineDiscussions || false
+                    always_divide_inline_discussions: alwaysDivideInlineDiscussions || false,
+                    available_division_schemes: availableDivisionSchemes || ['cohort']
                 };
             };
 
             createMockDiscussionsSettings = function(dividedInlineDiscussions,
                                                      dividedCourseWideDiscussions,
-                                                     alwaysDivideInlineDiscussions) {
+                                                     alwaysDivideInlineDiscussions,
+                                                     availableDivisionSchemes) {
                 return new CourseDiscussionsSettingsModel(
                     createMockDiscussionsSettingsJson(dividedInlineDiscussions,
                                                       dividedCourseWideDiscussions,
-                                                      alwaysDivideInlineDiscussions)
+                                                      alwaysDivideInlineDiscussions,
+                                                      availableDivisionSchemes)
                 );
             };
 
@@ -132,14 +136,14 @@ define(['backbone', 'jquery', 'edx-ui-toolkit/js/utils/spec-helpers/ajax-helpers
                 expect($courseWideDiscussionsForm.text()).
                     toContain('Course-Wide Discussion Topics');
                 expect($courseWideDiscussionsForm.text()).
-                    toContain('Select the course-wide discussion topics that you want to divide by cohort.');
+                    toContain('Select the course-wide discussion topics that you want to divide.');
 
                 // Should see the inline discussions form and its content
                 expect($inlineDiscussionsForm.length).toBe(1);
                 expect($inlineDiscussionsForm.text()).
                     toContain('Content-Specific Discussion Topics');
                 expect($inlineDiscussionsForm.text()).
-                    toContain('Specify whether content-specific discussion topics are divided by cohort.');
+                    toContain('Specify whether content-specific discussion topics are divided.');
             };
 
             beforeEach(function() {
