+++ lms/static/sass/course/instructor/_instructor_2.scss
@@ -1241,6 +1241,36 @@
 // --------------------
 .instructor-dashboard-wrapper-2 section.idash-section#discussions_management {
 
+  .division-scheme-container {
+    // See https://css-tricks.com/snippets/css/a-guide-to-flexbox/
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    .division-scheme {
+      font-size: 18px;
+    }
+
+    .division-scheme-item {
+      padding-left: 1%;
+      padding-right: 1%;
+      float: left;
+    }
+
+    .three-column-layout {
+       max-width: 33%;
+    }
+
+    .two-column-layout {
+       max-width: 50%;
+    }
+
+    .field-message {
+        font-size: 13px;
+    }
+  }
+
+  // cohort management
   .form-submit {
     @include idashbutton($uxpl-blue-base);
     @include font-size(14);
