+++ lms/templates/instructor/instructor_dashboard_2/divided-discussions-course-wide.underscore
@@ -1,11 +1,10 @@
-<h3 class="hd hd-3 subsection-title"><%- gettext('Specify whether discussion topics are divided by cohort') %></h3>
 <form action="" method="post" id="cohort-course-wide-discussions-form" class="cohort-course-wide-discussions-form">
     <div class="wrapper discussions-management-supplemental">
         <div class="form-fields">
             <div class="form-field">
                 <div class="course-wide-discussion-topics">
                     <h4 class="hd hd-4 subsection-title"><%- gettext('Course-Wide Discussion Topics') %></h4>
-                    <p><%- gettext('Select the course-wide discussion topics that you want to divide by cohort.') %></p>
+                    <p><%- gettext('Select the course-wide discussion topics that you want to divide.') %></p>
                     <div class="field">
                         <ul class="discussions-wrapper"><%= HtmlUtils.ensureHtml(courseWideTopicsHtml) %></ul>
                     </div>
